This is a anti theft device for CCTV Cameras based on Arduino And Raspberry Pi as an aggregation device
Internet of Things is all about connecting anything, anytime and
anywhere. With the advent of the government taking up initiatives like
Smart Cities, we as budding engineering students who are venturing into
the new world, wanted to do a project that is in line with the current
technology of the real world. We wanted to do a project that lets us
give back something to the University that is providing our education
and hence decided to do a project related to Smart Campus. Upon showing
a proposal to the Director (E&T), SRM University Kattankulathur, he
discussed his concern regarding the safety of CCTV cameras inside the
university campus and requested a project to be done regarding the same.
Hence, we decided to take it up as our minor project, leading to an
“Anti-theft device for CCTV cameras”. This project focuses on providing
just in time alerts to the concerned authorities upon the theft of the
CCTV cameras inside the campus.
Two levels of security are used in this device. The first level is a
wire that connects the CCTV Camera and the controller. If the wire is
not cut, the connection is close and current passes through the port to
which the CCTV Camera is connected, to the controller. If the wire is
cut, the connection is open and the current does not pass through the
port to which the CCTV camera is connected to the controller, this
happens when the CCTV camera and the anti-theft device are physically
separated from each.
The second level consists of two components, an IR Transmitter such as
an IR LED, and an IR receiver circuit. The IR LED and the receiver as a
combination can act as a proximity sensor. This combination is attached
to the camera and as long as the camera and the wall are in close
proximity, the camera is safe. When a person removes the camera, the
wall and the camera are no longer in close proximity. This triggers the
alerts.
In order to send the alerts, we are using a GSM module that is connected
to the Arduino to send SMS and make calls. The Raspberry Pi connects to
the Internet and sends email and posts status of the CCTV camera to the
online portal. Multiple Arduinos can be connected to a single Pi to
allow scalability for real world usage. All the cameras in a single
floor can have a common Raspberry Pi module through which they connect
to the Internet.